+++ tracer/bigsim/TraceReader.cpp
@@ -41,7 +41,7 @@ void TraceReader::loadOffsets(){
 }
 
 void TraceReader::readTrace(int* tot, int* totn, int* emPes, int* nwth, PE* pe,
-    int penum, int jobnum, unsigned long long* startTime)
+    int penum, int jobnum, double* startTime)
 {
   *nwth = numWth;
   *tot = totalWorkerProcs;
@@ -142,7 +142,7 @@ void TraceReader::readTrace(int* tot, int* totn, int* emPes, int* nwth, PE* pe,
 
 void TraceReader::setTaskFromLog(Task *t, BgTimeLog* bglog, int taskPE, int myEmPE, int jobPEindex, PE* pe)
 {
-  t->execTime=(unsigned long long)(((double)TIME_MULT * bglog->execTime));
+  t->execTime = (double)TIME_MULT * bglog->execTime;
   t->myMsgId.pe = bglog->msgId.pe() + jobPEindex;
   if(t->myMsgId.pe < 0)
     t->myMsgId.pe = -1;
@@ -163,7 +163,7 @@ void TraceReader::setTaskFromLog(Task *t, BgTimeLog* bglog, int taskPE, int myEm
     t->myEntries[i].thread = bglog->msgs[i]->tID;
 
     // sendTime is absolute
-    t->myEntries[i].sendOffset = (long long)(((double)TIME_MULT * (bglog->msgs[i]->sendTime - bglog->startTime)));
+    t->myEntries[i].sendOffset = (double)TIME_MULT * (bglog->msgs[i]->sendTime - bglog->startTime);
     assert(t->myEntries[i].sendOffset >= 0);
     t->myEntries[i].msgId.size = bglog->msgs[i]->msgsize;
   }
